+++ front/src/pages/Chat/interface/chanInfo.tsx
@@ -1,0 +1,6 @@
+interface ChanInfo {
+	channelName: string,
+	password: string
+}
+
+export default ChanInfo;
