+++ front/src/pages/Chat/interface/chanPassword.tsx
@@ -1,0 +1,6 @@
+interface ChanPassword {
+	oldPassword: string,
+	password: string
+}
+
+export default ChanPassword;
