+++ front/src/pages/Chat/interface/channelDto.tsx
@@ -1,0 +1,7 @@
+interface Channel {
+	id: number;
+	name: string;
+	statue: string;
+}
+
+export default Channel;
