+++ front/src/pages/Chat/interface/messageDto.tsx
@@ -1,0 +1,7 @@
+interface Message {
+	message: string;
+	username: string;
+	uId: number
+}
+
+export default Message;
